+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds metrics patch for discord.py's HTTP requests.
+
 ## [v7.10.9](https://github.com/lexicalunit/spellbot/releases/tag/v7.10.9) - 2021-12-27
 
 ### Added